requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Dropzone==1.5.4
google-api-core==1.17.0
google-auth==1.14.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
scipy==1.5.2
six==1.15.0
tqdm==4.47.0
# urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1